Занялся загрузкой объектов форума

+++ app/src/main/java/com/example/learntogether/FromAPI/InfoBase.java
@@ -3,5 +3,5 @@ package com.example.learntogether.FromAPI;
 public class InfoBase {
     public int ID_InfoBase, ID_Author, ID_Group, CommentsFound;
     public String Title, Text, Type, WhenAdd, AuthorTitle;
-    public float Rate = 5;
+    public double Rate = 5;
 }
